refactor(server): Add tower pipeline base to server
